tcp: Use PktBlock->TSFirst for SnapshotTS on both Flow and TCP
This lets us ensure that the "sync point" for the TCPEvents and their respective
FlowRecord's can be managed per SnapshotTS.

This needs to be TSFirst, instead of TSSnapshot because that's only set later
when a "snapshot" block is created (see: s_FlowSampleTSLast) - we need the same
SnapshotTS to be used for every TCPEvent and for the FlowRecord's they were
generated for.
TSFirst offers that.